Persistence achieved.
Changes:
* Added utils to module exports and exposed Gson. This is necessary for
persistent use cases.
* Added @Exclude for strategic disabling of (de)serialization
* Added deserialization to Population (instead of replacing it with a
List entirely). Population is needed because it is more semantically
meaningful as part of a whole population or partial population (i.e.,
species)
* Innovations are now tracked with 64-bit longs
* Refactored repopulation in SpeciatedPopulation, and enabled
specification of details wiped out by JSON serialization
* Added various file manipulations (as opposed to solely resource
manipulations) to ResourceUtils. Files are useful resources :)
* Added PersistenceTest, which learns 5 generations, starting with
generation 100 (which is persisted in a file)
* Made sure to reset selector(s) after a generation has passed
* Mild formatting changes

+++ src/test/java/test/NeuralIndividualTest.java
@@ -141,7 +141,7 @@ public class NeuralIndividualTest
 		Assertions.assertEquals(0, ind.genome().genes().size());
 	}
 
-	private boolean inRange(int num, int low, int high)
+	private boolean inRange(long num, long low, long high)
 	{
 		return low <= num && num <= high;
 	}
